adds lightning support, better handling of metrics
