+++ mantisshrimp/engines/lightning/__init__.py
@@ -1,0 +1,2 @@
+from mantisshrimp.engines.lightning.lightning_model_adapter import *
+from mantisshrimp.engines.lightning.rcnn_lightning_model import *
