+++ mantisshrimp/engines/lightning/rcnn_lightning_model.py
@@ -1,0 +1,40 @@
+__all__ = ["RCNNLightningModel"]
+
+from mantisshrimp.imports import *
+from mantisshrimp.utils import *
+from mantisshrimp.models import *
+from mantisshrimp.metrics import *
+from mantisshrimp.engines.lightning.lightning_model_adapter import LightningModelAdapter
+
+
+class RCNNLightningModel(LightningModelAdapter, ABC):
+    def __init__(self, model: MantisRCNN, metrics: List[Metric]):
+        super().__init__(metrics=metrics)
+        self.model = model
+
+    def forward(self, *args, **kwargs):
+        return self.model(*args, **kwargs)
+
+    def training_step(self, batch, batch_idx):
+        xb, yb = batch
+        preds = self(xb, yb)
+        loss = self.model.loss(preds, yb)
+        log = {"train/loss": loss}
+        return {"loss": loss, "log": log}
+
+    def validation_step(self, batch, batch_idx):
+        xb, yb = batch
+        with torch.no_grad():
+            self.train()
+            preds = self(xb, yb)
+            loss = self.model.loss(preds, yb)
+            self.eval()
+            preds = self(xb)
+            self.accumulate_metrics(xb, yb, preds)
+        return {"valid/loss": loss}
+
+    def validation_epoch_end(self, outs):
+        loss_log = {k: torch.stack(v).mean() for k, v in mergeds(outs).items()}
+        metrics_log = self.finalize_metrics()
+        log = {**loss_log, **metrics_log}
+        return {"val_loss": log["valid/loss"], "log": log}
