+++ mantisshrimp/metrics/coco_metric/coco_metric.py
@@ -19,18 +19,25 @@ class COCOMetric(Metric):
         if keypoint:
             self.iou_types.append("keypoints")
 
+        self.reset()
+
     def reset(self):
-        self.coco_evaluator = CocoEvaluator(self._coco_ds, self.iou_types)
+        self._coco_evaluator = CocoEvaluator(self._coco_ds, self.iou_types)
 
     def accumulate(self, xb, yb, preds):
         # TODO: Implement batch_to_cpu helper function
         assert len(xb) == len(yb) == len(preds)
         preds = [{k: v.to(torch.device("cpu")) for k, v in p.items()} for p in preds]
         res = {y["image_id"].item(): pred for y, pred in zip(yb, preds)}
-        self.coco_evaluator.update(res)
+        self._coco_evaluator.update(res)
 
-    def finalize(self) -> float:
-        self.coco_evaluator.synchronize_between_processes()
-        self.coco_evaluator.accumulate()
-        self.coco_evaluator.summarize()
-        return 0
+    def finalize(self) -> Dict[str, float]:
+        self._coco_evaluator.synchronize_between_processes()
+        self._coco_evaluator.accumulate()
+        self._coco_evaluator.summarize()
+        logs = {
+            f"{k}_ap": float(v.stats[0])
+            for k, v in self._coco_evaluator.coco_eval.items()
+        }
+        self.reset()
+        return logs
